[BBC-9128] Drawer
